- Trying to implement model skeletal animation. Added yet empty classes that will form a animation.  + Added BoneTransform class and represent a transformation of bone  + Added Bone class that represent a bone from model  + Added KeyFrame class. It's represent a frame from animation  + Added Animation class. Contains all frames of animation.  + Added AnimatorLoader class and it loading all data from file

+++ 3DGame/player.h
@@ -17,7 +17,7 @@ private:
 
 	//coords
 	int tileX, tileY;
-	int worldOriginX, worldOriginZ;
+	float worldOriginX, worldOriginZ;
 	Model model;
 
 	Player();
